refactor: extract temperature helpers
